Python/requirements.txt
mysql-connector-python>=8.0.0
paho-mqtt>=1.6.0
zhipuai>=2.0.0
